Integrate test refactors
